+++ xknx/knxip/__init__.py
@@ -24,10 +24,13 @@ from .knxip_enum import (
     HostProtocol,
     KNXIPServiceType,
     KNXMedium,
+    SearchRequestParameterType,
 )
 from .routing_indication import RoutingIndication
 from .search_request import SearchRequest
+from .search_request_extended import SearchRequestExtended
 from .search_response import SearchResponse
+from .search_response_extended import SearchResponseExtended
 from .secure_wrapper import SecureWrapper
 from .session_authenticate import SessionAuthenticate
 from .session_request import SessionRequest
@@ -66,7 +69,10 @@ __all__ = [
     "KNXMedium",
     "RoutingIndication",
     "SearchRequest",
+    "SearchRequestExtended",
+    "SearchRequestParameterType",
     "SearchResponse",
+    "SearchResponseExtended",
     "SecureWrapper",
     "SessionAuthenticate",
     "SessionRequest",